Re: [PATCH 5.6.1] test suite Message-ID: <20010624054705.A27890@math.ohio-state.edu>
Test suite fixes.

p4raw-id: //depot/perl@10883

+++ ext/POSIX/sigaction.t
@@ -118,10 +118,12 @@ eval {
 print $@ ? "ok 17\n" : "not ok 17\n";
 
 $newaction=POSIX::SigAction->new(sub { $ok10=1; });
-sigaction(SIGCONT, POSIX::SigAction->new('DEFAULT'));
-{
+if (eval { SIGCONT; 1 }) {
+    sigaction(SIGCONT, POSIX::SigAction->new('DEFAULT'));
+    {
 	local($^W)=0;
 	kill 'CONT', $$;
+    }
 }
 print $bad18 ? "not ok 18\n" : "ok 18\n";
